pyproject.toml
[project]
name = "content-pilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
	"gspread>=6.2.1",
	"langchain-community>=0.4.1",
	"langchain-tavily>=0.2.17",
	"langchain[openai]>=0.3.27",
	"langgraph>=0.6.7",
	"langgraph-cli[inmem]>=0.4.2",
	"langsmith>=0.4.28",
	"oauth2client>=4.1.3",
	"openai[aiohttp]>=1.107.3",
	"playwright>=1.58.0",
	"python-dotenv>=1.1.1",
	"tavily-python>=0.7.23",
	"tiktoken>=0.11.0",
]

[dependency-groups]
dev = [
  "ipykernel>=6.30.1",
]
